ready to migrate

+++ app/models/activity.py
@@ -7,8 +7,8 @@ class Activity(db.Model):
   id = db.Column(db.Integer, primary_key = True)
   title = db.Column(db.String(255), nullable = False)
   description = db.Column(db.String(2000))
-  distance = db.Column(db.Integer, nullable = False)
-  time = db.Column(db.Integer, nullable = False)
+  distance = db.Column(db.Float, nullable = False)
+  time = db.Column(db.Float, nullable = False)
   gpx_file = db.Column(db.String(255))
   user_id = db.Column(db.Integer, nullable = False, db.ForeignKey('users.id'))
